使用md5校验文件内容，上传成统一名称

+++ app/upload.class.php
@@ -9,9 +9,9 @@ class Upload {
         //$file_type = $file["type"];
         $file_type = explode('.', $file['name']);
         $file_type = $file_type[1];
-        $new_name = md5($file['size'] . $file['name']) . '.' . $file_type;
+
         logging::d('upload_image', 'file_type:' . $file_type);
-        logging::d('upload_image', 'new_name:' . $new_name);
+
         
         if(!is_uploaded_file($file['tmp_name'])){
           //如果不是通过HTTP POST上传的
@@ -27,6 +27,13 @@ class Upload {
             }
         }
         
+        $img_src = file_get_contents($file['tmp_name']);
+        //logging::d("UPLOAD", "imgsrc:" . $img_src);
+        $new_name = md5($img_src) . '.' . strtolower($file_type);
+        if(file_exists(UPLOAD_DIR . "/" . $new_name)){
+            return $new_name;
+        }
+        logging::d('upload_image', 'new_name:' . $new_name);
         if(!move_uploaded_file($file['tmp_name'], UPLOAD_DIR . "/" . $new_name)){
             logging::d('upload_image', 'move_uploaded:' . 'move_uploaded失败');
             return false;
